// src/graphql/queries/me.graphql.tsx
export type Maybe<T> = T | null;
// ...
export type AggregateSession = {
  __typename?: 'AggregateSession';
  count?: Maybe<SessionCountAggregate>;
  min?: Maybe<SessionMinAggregate>;
  max?: Maybe<SessionMaxAggregate>;
};

export type AggregateTokens = {
  __typename?: 'AggregateTokens';
  count?: Maybe<TokensCountAggregate>;
  min?: Maybe<TokensMinAggregate>;
  max?: Maybe<TokensMaxAggregate>;
};

export type AggregateUser = {
  __typename?: 'AggregateUser';
  count?: Maybe<UserCountAggregate>;
  avg?: Maybe<UserAvgAggregate>;
  sum?: Maybe<UserSumAggregate>;
  min?: Maybe<UserMinAggregate>;
  max?: Maybe<UserMaxAggregate>;
};
// ...
export type FieldError = {
  __typename?: 'FieldError';
  field: Scalars['String'];
  message: Scalars['String'];
};
// ...
export type Query = {
  __typename?: 'Query';
  session?: Maybe<Session>;
  findFirstSession?: Maybe<Session>;
  sessions: Array<Session>;
  aggregateSession: AggregateSession;
  groupBySession: Array<SessionGroupBy>;
  findUniqueTokens?: Maybe<Tokens>;
  findFirstTokens?: Maybe<Tokens>;
  findManyTokens: Array<Tokens>;
  aggregateTokens: AggregateTokens;
  groupByTokens: Array<TokensGroupBy>;
  user?: Maybe<User>;
  findFirstUser?: Maybe<User>;
  users: Array<User>;
  aggregateUser: AggregateUser;
  groupByUser: Array<UserGroupBy>;
  me: UserResponse;
};
// ...
export type Session = {
  __typename?: 'Session';
  sid: Scalars['String'];
  sess: Scalars['JSON'];
  expire: Scalars['DateTime'];
};
// ...
export type SessionGroupBy = {
  __typename?: 'SessionGroupBy';
  sid: Scalars['String'];
  sess: Scalars['JSON'];
  expire: Scalars['DateTime'];
  count?: Maybe<SessionCountAggregate>;
  min?: Maybe<SessionMinAggregate>;
  max?: Maybe<SessionMaxAggregate>;
};
// ...
export type Tokens = {
  __typename?: 'Tokens';
  id: Scalars['String'];
  type: TokenType;
  userId: Scalars['String'];
  expire: Scalars['DateTime'];
  createdAt: Scalars['DateTime'];
  token: Scalars['String'];
};
// ...
export type TokensGroupBy = {
  __typename?: 'TokensGroupBy';
  id: Scalars['String'];
  type: TokenType;
  userId: Scalars['String'];
  expire: Scalars['DateTime'];
  createdAt: Scalars['DateTime'];
  token: Scalars['String'];
  count?: Maybe<TokensCountAggregate>;
  min?: Maybe<TokensMinAggregate>;
  max?: Maybe<TokensMaxAggregate>;
};
// ...
export type User = {
  __typename?: 'User';
  id: Scalars['String'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name?: Maybe<Scalars['String']>;
  role: UserRole;
  email: Scalars['String'];
  verified: Scalars['Boolean'];
  stripeId?: Maybe<Scalars['String']>;
  coins: Scalars['BigInt'];
  ccLast4?: Maybe<Scalars['Int']>;
  patronageType?: Maybe<Patronage>;
};
// ...
export type UserGroupBy = {
  __typename?: 'UserGroupBy';
  id: Scalars['String'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name?: Maybe<Scalars['String']>;
  role: UserRole;
  email: Scalars['String'];
  password: Scalars['String'];
  verified: Scalars['Boolean'];
  stripeId?: Maybe<Scalars['String']>;
  coins: Scalars['BigInt'];
  ccLast4?: Maybe<Scalars['Int']>;
  patronageType?: Maybe<Patronage>;
  count?: Maybe<UserCountAggregate>;
  avg?: Maybe<UserAvgAggregate>;
  sum?: Maybe<UserSumAggregate>;
  min?: Maybe<UserMinAggregate>;
  max?: Maybe<UserMaxAggregate>;
};
// ...
export type UserResponse = {
  __typename?: 'UserResponse';
  errors?: Maybe<Array<FieldError>>;
  user?: Maybe<User>;
};
// ...
export type MeQuery = (
  { __typename?: 'Query' }
  & { me: (
    { __typename: 'UserResponse' }
    & RegularUserResponseFragment
  ) }
);

export type RegularUserResponseFragment = (
  { __typename?: 'UserResponse' }
  & { errors?: Maybe<Array<(
    { __typename?: 'FieldError' }
    & RegularErrorFragment
  )>>, user?: Maybe<(
    { __typename?: 'User' }
    & RegularUserFragment
  )> }
);

export type RegularErrorFragment = (
  { __typename: 'FieldError' }
  & Pick<FieldError, 'field' | 'message'>
);

export type RegularUserFragment = (
  { __typename: 'User' }
  & Pick<User, 'id' | 'name' | 'email'>
);

export const RegularErrorFragmentDoc = gql`
    fragment RegularError on FieldError {
  field
  message
  __typename
}
    `;
export const RegularUserFragmentDoc = gql`
    fragment RegularUser on User {
  id
  name
  email
  __typename
}
    `;
export const RegularUserResponseFragmentDoc = gql`
    fragment RegularUserResponse on UserResponse {
  errors {
    ...RegularError
  }
  user {
    ...RegularUser
  }
}
    ${RegularErrorFragmentDoc}
${RegularUserFragmentDoc}`;
export const MeDocument = gql`
    query Me {
  me {
    ...RegularUserResponse
    __typename
  }
}
    ${RegularUserResponseFragmentDoc}`;